+++ tests/conftest.py
@@ -1,0 +1,4 @@
+import pytest # noqa # pylint: disable=unused-import
+
+# Shared fixtures
+from .test_tracing import collector # noqa # pylint: disable=unused-import