+++ tests/testmodule/__init__.py
@@ -1,0 +1,4 @@
+class Foo:
+    def __init__(self, arg1: str, arg2: int) -> None:
+        self.arg1 = arg1
+        self.arg2 = arg2